Add token authentication support for google home api
Since June 2019 the Google Home Local API requires authentication using
a "local authorization token" for most requests. This commit changes the
affected api requests to send a user-provided authentication token and
switch to the now used port 8443 for the api.

A special case is the eureka_info endpoint, that can still be accessed
over the old port without authentication, but provides more information,
when accessed over an encrypted and authorized connection. The method
thus supports calling with AND without a token.